requirements.txt
beautifulsoup4==4.12.2
numpy==1.25.2
faiss-cpu==1.7.4
transformers==4.34.0
sentence-transformers==2.2.2
requests==2.31.0
google-generativeai==0.2.0
dotenv==0.19.0
